backend/requirements.txt
# Flask
Flask==3.1.0
Flask-Mail==0.10.1
gunicorn
python-dotenv
flask-cors
